refactor & add async structure in ckpt_schedule
