+++ megatron/spiral/schedules.py
@@ -2,8 +2,7 @@ import contextlib
 import warnings
 import nvtx
 import sys
-from typing import Callable, Iterator, List, Optional, Union, Tuple
-from enum import Enum
+from typing import Callable, Iterator, List, Optional, Union
 from collections import deque
 from queue import Queue
 import threading
@@ -11,7 +10,7 @@ import threading
 import torch
 from torch.nn.parallel.distributed import DistributedDataParallel as torchDDP
 
-from megatron import get_num_microbatches, get_args, get_timers
+from megatron import get_args, get_timers
 from megatron.core import mpu
 from megatron.core.enums import ModelType
 from megatron.core.utils import get_model_type, get_attr_wrapped_model
@@ -23,140 +22,14 @@ from megatron.spiral.init_context import SpiralParamStatus
 from megatron.spiral.utils import is_spiral_param
 from megatron.spiral.generic import ContextManagers
 import megatron.spiral.build_state as sbs
-
-
-# Due to Tuple() ctor support from python >= 3.9
-_PYTHON_VERSION = sys.version_info
+from megatron.spiral.ckpt_communication import comm_input_ckpt
+from megatron.spiral.ckpt_schedule import CkptSendRecvType, CkptSendRecvSchedule, CkptSendRecvOp
 
 # Types
 Shape = Union[List[int], torch.Size]
 
 # Constants
 _DEBUG_SCHEDULE = True
-
-class CkptSendRecvType(Enum):
-    SEND = "send"
-    RECV = "recv"
-
-
-CkptSendRecvOp = Tuple[CkptSendRecvType, int, int]  # (comm_type, phase_id, rank)
-
-
-class CkptSendRecvSchedule:
-    """A schedule for sending and receiving input tensor checkpoints.
-
-    Construct global schedule for all timestep of all rank.
-    A timestep is a time unit for a micro-batch fwd/bwd computation.
-    All timestep means from beginning of the minibatch on PP rank 0 to the end of the minibatch on PP rank N-1.
-    Thus, a rank's schedule includes non_compute timesteps where no fwd/bwd computation is performed due to pipeline fill/drain.
-    non_compute timesteps are blank timesteps in below example
-
-    Example:
-    _SPIRAL_FORWARD_VIRTUAL_SIZE = 3
-    _SPIRAL_BACKWARD_VIRTUAL_SIZE = 3
-    _MPU_PIPELINE_MODEL_PARALLEL_WORLD_SIZE = 4
-
-        | ts0   | ts1   | ts1   | ts1   | ts1   | ts1   | ts1   | ts1   | ts1   | ts1   | ts1   |
-    r0  | fstart| comp  | comp  | comp  | bstart| comp  | comp  | comp  |       |       |       |
-    r1  |       | comp  | comp  | comp  | comp  | comp  | comp  | comp  | comp  |       |       |
-    r2  |       |       | comp  | comp  | comp  | comp  | comp  | comp  | comp  | comp  |       |
-    r3  |       |       |       | comp  | comp  | comp  | fend  | comp  | comp  | comp  | bend  |
-
-    """
-
-    def __init__(self, num_microbatches: Optional[int]):
-        if num_microbatches is None:
-            num_microbatches = get_num_microbatches()
-        assert num_microbatches >= mpu.get_pipeline_model_parallel_world_size(
-        ),  "CkptSendRecvSchedule requires num_microbatches >= pipeline model parallel world size"
-        self.num_microbatches = num_microbatches
-
-        num_compute_ts = (
-            mpu.get_spiral_forward_virtual_size()
-            + mpu.get_spiral_backward_virtual_size()
-        ) * self.num_microbatches
-        num_non_compute_ts = mpu.get_pipeline_model_parallel_world_size() - 1
-        self.total_ts = num_compute_ts + num_non_compute_ts
-        self.global_schedule: List[List[List[CkptSendRecvOp]]] = [
-            [[] for _ in range(self.total_ts)]
-            for _ in range(mpu.get_pipeline_model_parallel_world_size())
-        ]
-
-        self._set_recv_schedule()
-        self._set_send_schedule()
-        self._schedule_generator = self._generator()
-
-    def _set_recv_schedule(self):
-        for pp_rank in range(mpu.get_pipeline_model_parallel_world_size()):
-            num_pre_pipeline_non_compute_ts = pp_rank
-            recv_start_ts = (
-                num_pre_pipeline_non_compute_ts
-                + mpu.get_spiral_forward_virtual_size() * self.num_microbatches
-            )
-            curr_ts = recv_start_ts
-
-            for bwd_stage_id in range(
-                mpu.get_spiral_backward_virtual_size() - 1, -1, -1
-            ):
-                bwd_phases_start = (
-                    sbs.get_spiral_backward_stage_build_phase_size()
-                    * mpu.get_pipeline_model_parallel_world_size()
-                    * bwd_stage_id
-                    + sbs.get_spiral_backward_stage_build_phase_size()
-                    * (mpu.get_pipeline_model_parallel_world_size() - pp_rank - 1)
-                )
-                recv_phase = bwd_phases_start  # modify when needed
-                for _ in range(self.num_microbatches):
-                    if _PYTHON_VERSION >= (3, 9):
-                        _op = CkptSendRecvOp(
-                            CkptSendRecvType.RECV,
-                            recv_phase,
-                            sbs.get_pp_rank_for_fwd_phase(recv_phase),
-                        )
-                    else:
-                        _op = (
-                            CkptSendRecvType.RECV,
-                            recv_phase,
-                            sbs.get_pp_rank_for_fwd_phase(recv_phase),
-                        )
-                    self.global_schedule[pp_rank][curr_ts].append(_op)
-                    curr_ts += 1
-
-    def _set_send_schedule(self):
-        # TODO (SpiralPipe) sends of only self is required
-        for j in range(self.total_ts):
-            for pp_rank in range(mpu.get_pipeline_model_parallel_world_size()):
-                for recv_comm in filter(
-                    lambda x: x[0] == CkptSendRecvType.RECV,
-                    self.global_schedule[pp_rank][j],
-                ):
-                    # add send to dst for the sender rank queue
-                    _src_rank = recv_comm[2]
-                    if _PYTHON_VERSION >= (3, 9):
-                        _op = CkptSendRecvOp(
-                            CkptSendRecvType.SEND, recv_comm[1], pp_rank
-                        )
-                    else:
-                        _op = (CkptSendRecvType.SEND, recv_comm[1], pp_rank)
-                    self.global_schedule[_src_rank][j].append(_op)
-
-    def __str__(self) -> str:
-        _str = ""
-        for pp_rank in range(len(self.global_schedule)):
-            _str += f"rank {pp_rank}\n"
-            for ts, comms in enumerate(self.global_schedule[pp_rank]):
-                _str += f"\tts {ts}: {comms}\n"
-        return _str
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        return next(self._schedule_generator)
-
-    def _generator(self):
-        for comms in self.global_schedule[mpu.get_pipeline_model_parallel_rank()]:
-            yield comms
 
 
 def _post_wait_set_spiral_param_status(module, status: SpiralParamStatus):
@@ -267,119 +140,6 @@ def forward_backward_pipelining_with_spiral_remap(
         grad_scaler = kwargs["spiral_grad_scaler"]
     if optimize_after_bwd_stage:
         optimizer_threads = [] # (thread, queue, ev)
-
-    # TODO (SpiralPipe) Move to spiral_p2p
-    @nvtx.annotate("comm_input_ckpt", color="red")
-    def comm_input_ckpt(schedule):
-        if _DEBUG_SCHEDULE:
-            spiral_print(f"comm: {schedule}")
-
-        tensor_sends = []
-        send_ranks = []
-        recv_ranks = []
-
-        insert_idx_to_recvs = []
-        insert_value_to_recvs = []
-
-        recv_idx = 0
-        send_idx = 0
-        for idx, (comm_type, phase_id, rank) in enumerate(schedule):
-
-            phase_fwd_rank = sbs.get_pp_rank_for_fwd_phase(phase_id)
-            local_stage_id, local_phase_id = sbs.fwd_phase2local_stage_phase(phase_id)
-
-            _prefix = str(schedule[idx])
-
-            if comm_type == CkptSendRecvType.RECV:
-                assert (
-                    phase_fwd_rank == rank
-                ), f"[RECV] phase_fwd_rank = {phase_fwd_rank}, rank = {rank} mismatch"
-
-                if phase_id == 0:
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(_prefix + " Phase 0 => insert None to recvs")
-                    insert_idx_to_recvs.append(recv_idx)
-                    insert_value_to_recvs.append(None)
-                elif rank == mpu.get_pipeline_model_parallel_rank():
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(_prefix + " Self recv => pop ckpt and insert to recvs")
-                    insert_idx_to_recvs.append(recv_idx)
-                    # NOTE (SpiralPipe) Using the popped input tensor from original fwd stage can lead to trouble, as it contains
-                    # the computation graph constructed already. We are prone to this error when #fwd != #bwd and hence the same rank
-                    # can perform fwd and bwd of the same phase. Re-computation using this tensor will lead to duplicated computation
-                    # graph being constructed. So, we currently perform the original FWD in torch.no_grad() mode, and then recompute
-                    # in BWD without torch.no_grad(). Another solution may exist.
-                    input_ckpt_ = (
-                        model[local_stage_id]
-                        .module[local_phase_id]
-                        .spiral_input_tensors.popleft()
-                        .detach()
-                        .requires_grad_()
-                    )
-                    assert (
-                        input_ckpt_.requires_grad
-                    ), "Input ckpt must require grad before feeding to BWD"
-                    insert_value_to_recvs.append(input_ckpt_)
-                else:
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(
-                            _prefix + " Recv from other rank => append to recv_ranks"
-                        )
-                    recv_ranks.append(rank)
-                recv_idx += 1
-
-            elif comm_type == CkptSendRecvType.SEND:
-                assert (
-                    phase_fwd_rank == mpu.get_pipeline_model_parallel_rank()
-                ), f"[SEND] phase_fwd_rank = {phase_fwd_rank}, self = {mpu.get_pipeline_model_parallel_rank()} mismatch"
-
-                if phase_id == 0:
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(_prefix + " Phase 0 => skip")
-                elif rank == mpu.get_pipeline_model_parallel_rank():
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(_prefix + " Self send => skip")
-                else:
-                    if _DEBUG_SCHEDULE:
-                        spiral_print(
-                            _prefix
-                            + " Send to other rank => pop ckpt & append to tensor sends and append to send_ranks"
-                        )
-                    tensor_sends.append(
-                        model[local_stage_id]
-                        .module[local_phase_id]
-                        .spiral_input_tensors.popleft()
-                    )
-                    send_ranks.append(rank)
-                send_idx += 1
-            else:
-                raise RuntimeError(f"Invalid comm type {comm_type}")
-
-        recvs, reqs = None, []  # placeholder
-        if len(send_ranks) > 0 or len(recv_ranks) > 0:
-            if mpu.is_spiral_cross_mapping():
-                # translate cm_rank back to pp_rank before communication
-                send_ranks = [mpu.translate_cm_rank_to_pp_rank(rank) for rank in send_ranks]
-                recv_ranks = [mpu.translate_cm_rank_to_pp_rank(rank) for rank in recv_ranks]
-
-            recvs, reqs = spiral_p2p._communicate(
-                tensor_sends=tensor_sends if len(tensor_sends) > 0 else None,
-                send_ranks=send_ranks if len(send_ranks) > 0 else None,
-                recv_ranks=recv_ranks if len(recv_ranks) > 0 else None,
-                tensor_shape=tensor_shape,
-                group=mpu.get_spiral_input_tensor_ckpt_group(),
-                batch_p2p_comm=True,
-                wait_on_reqs=False,
-                dtype=dtype,
-            )
-
-        for recv_idx, recv_val in zip(insert_idx_to_recvs, insert_value_to_recvs):
-            if recvs is None:
-                recvs = []
-            recvs.insert(recv_idx, recv_val)
-
-        return recvs, reqs
-    # end comm_input_ckpt()
 
     def _cleanup():
         # cleanup checkpointed input tensors
@@ -500,7 +260,7 @@ def forward_backward_pipelining_with_spiral_remap(
 
             # send input tensor ckpt
             if not forward_only:
-                comm_input_ckpt(next(ckpt_send_recv_schedule))
+                comm_input_ckpt(next(ckpt_send_recv_schedule), model, tensor_shape, dtype)
 
             # set input tensor
             if mpu.is_pipeline_first_stage():
@@ -681,7 +441,7 @@ def forward_backward_pipelining_with_spiral_remap(
 
             # recv input tensor ckpt
             recv_input_ckpts, recv_input_ckpt_handle = comm_input_ckpt(
-                next(ckpt_send_recv_schedule)
+                next(ckpt_send_recv_schedule), model, tensor_shape, dtype
             )
 
             # set output tensor grad
@@ -893,7 +653,7 @@ def forward_backward_pipelining_with_spiral_remap(
     )
     for _ in range(__num_post_pipeline_non_compute_ts):
         if not forward_only:
-            comm_input_ckpt(next(ckpt_send_recv_schedule))
+            comm_input_ckpt(next(ckpt_send_recv_schedule), model, tensor_shape, dtype)
 
     # cleanup schedule events
     if (
